refactor(pathsafe): consolidate symlink-safe path check into one package
The InsideDir + deepestExistingAncestor logic lived in BOTH internal/config
and internal/prompts as separate copies — and the drift between them is what
caused the v0.17.1 disclosure bug (config hardened in v0.10.5, prompts copy
stayed lexical-only). A security primitive must not be duplicated where the
copies can drift.

Extract internal/pathsafe.InsideDir as the single source of truth; config and
prompts now both call it. Core path/symlink tests live in pathsafe_test.go;
prompts keeps the end-to-end TestResolve_SymlinkOverrideDoesNotLeak. Also
resolves the SonarCloud "duplication on new code" gate the ported copy tripped.
