+++ CLAUDE.md
@@ -107,7 +107,7 @@ These are things you can't infer by reading the code in 30 seconds:
 - **Audit uses the same probe + invoker path as review (v0.10+).** `local-review audit --topic <id>` picks ONE authenticated LLM (first match; claude when available) and walks `git ls-files`-grouped chunks through it. Chunks above the per-chunk cap (default 96 KiB) auto-split into `pkg [part N/M]` sub-chunks via greedy bin-packing in `internal/audit/walker.go`, preserving per-file adjacency. Negative `--max-bytes-per-chunk` is rejected up front.
 - **`isLocalURL` widened to RFC1918 + IPv6 ULA/link-local (v0.10.4), then CGNAT/Tailscale (v0.12.1).** Ollama on a LAN host (`http://192.168.x.x:11434/v1`) or over Tailscale (`http://100.x.x.x:11434/v1`, RFC 6598 `100.64.0.0/10`) no longer needs a dummy `api_key`. The Tailscale case (added after the Ollama-over-Tailscale dogfood) isn't covered by Go's `IsPrivate()`, so it's a separate `cgnatRange.Contains(ip)` check. Corporate-gateway invariant preserved via the existing `c.APIKey == ""` guard — bypass only fires when no key is configured. IPv6 zone suffixes (`fe80::1%en0`) stripped before `net.ParseIP`. See `internal/llm/client.go`.
 - **Language packs DON'T carry the JSON schema; it's appended at resolve time (v0.12.1).** Pre-fix every non-default pack ended with "Same JSON shape as the default pack" but `prompts.Resolve` only ever sent ONE pack, so a single-LLM review of any language never received the actual schema — strong cloud models inferred it, weak local models (Ollama) returned JSON without the `findings` key and the review failed to parse. Now `prompts.FindingsJSONSchema` is the single source of truth, appended by `Resolve` ONLY when `ResolveOptions.RequireJSON` is set. Historically the single-LLM-fallback path (`internal/review.Reviewer.Run`, removed in v0.15) set `RequireJSON=true`; the multi-LLM CLI invokers leave it false because they append a "respond in markdown, NOT JSON" override (a competing JSON schema there risks a stray JSON reply the merger can't read). With the v0.15 removal of the single-LLM path, `RequireJSON` no longer has an active caller in production but the flag survives in `ResolveOptions` for future re-introduction (e.g. an opt-in structured-JSON multi-LLM mode). Surfaced by the Ollama-over-Tailscale dogfood.
-- **`pathInsideDir` resolves symlinks before containment (v0.10.5).** Lexical-only check (v0.10.0) admitted symlink-escape paths. v0.10.5 adds `EvalSymlinks` + deepest-existing-ancestor walk-up (closes the missing-leaf bypass — `evil-link/new-leaf` where `evil-link → /etc`); fail-closed on resolve errors. See `internal/config/config.go`.
+- **Symlink-safe path containment lives in ONE place: `internal/pathsafe.InsideDir` (v0.17.1).** Lexical-only checks (v0.10.0) admitted symlink-escape paths; v0.10.5 added `EvalSymlinks` + deepest-existing-ancestor walk-up (closes the missing-leaf bypass — `evil-link/new-leaf` where `evil-link → /etc`), fail-closed on resolve errors. **This check was duplicated in `internal/config` AND `internal/prompts`, and the copies drifted: config got the v0.10.5 hardening, the prompts copy stayed lexical-only → a `major` arbitrary-file-disclosure vuln (`pack_dir/<lang>.md → /etc/passwd` leaked into the LLM prompt), found by the tool's own `audit --topic security` and fixed in v0.17.1.** Both packages now call `pathsafe.InsideDir` — a security primitive must not be duplicated where copies can drift. Don't re-inline it.
 - **The repo-level `.local-review.yml` is an UNTRUSTED config layer.** It's attacker-controllable when you review code you didn't write (CI checking out a hostile commit, a fresh clone). `config.mergeFrom` takes a `trusted bool`; `Load` passes `true` for the user-home layer and `false` for the repo layer (unless `LOCAL_REVIEW_TRUST_REPO_CONFIG=1`). `sanitizeUntrustedLayer` strips the security-sensitive LLM fields from the untrusted layer before merge — `cli_path` (→ `exec.LookPath`+`exec.CommandContext` = arbitrary RCE), `base_url` (→ a provider agent that POSTs the diff/source to that endpoint = exfiltration), and `api_key` (secret-in-repo) — each with a stderr warning. Non-sensitive fields (model/timeout/enabled/prompts/review) still merge. This is the same trust boundary `resolveRelativePaths` already drew for `prompts.pack_dir`, extended to the exec/network/secret fields. Closes the two `major` findings from the v0.15.1 L6 audit.
 - **A name with `base_url` set is a provider agent, never also a CLI one (`dropCLITwins`, `cmd/local-review/runner.go`).** Pre-fix, `llms.claude.base_url: …` with the claude CLI installed produced BOTH a CLI `claude` and a provider `claude` — two same-named agents that double-reviewed and collided in the name-keyed ready/merge maps. `pickAgents` now drops the CLI twin (BaseURL=="") of any config name that carries a `base_url`, so the provider entry wins.
 - **Antigravity (`agy`) is DETECTED but NOT a reviewer (2026-05).** Google's Gemini-CLI successor is surfaced by `doctor` (`◐ experimental` row) but gated out of the fan-out by `cli.IsReviewCapable("antigravity") == false`. The authenticated dogfood showed agy's headless `--print` runs a full autonomous agent loop: it explores the repo, reconstructs its OWN diff (ignoring the one passed), and streams tool-step narration ("I will run git diff…") to stdout instead of a review — a real `review --only antigravity` yielded 6.5 KB of narration, zero findings, empty merge. Short prompts can also return empty stdout. `AntigravityInvoker` exists as scaffolding only (reachable from `NewInvoker` for type/interface tests, never a real run). `--only antigravity` refuses cleanly. Don't wire it into the active set without first solving the structured-output / suppress-agent-loop invocation contract. Gemini stays the live Google agent until its 2026-06-18 cutoff.
